Refactor ppo2.py (move shared code elsewhere)
