+++ stable_baselines/sac/sac.py
@@ -9,8 +9,9 @@ import tensorflow as tf
 from stable_baselines.common import tf_util, OffPolicyRLModel, SetVerbosity, TensorboardWriter
 from stable_baselines.common.vec_env import VecEnv
 from stable_baselines.common.tf_util import total_episode_reward_logger
+from stable_baselines.common.math_util import safe_mean
+from stable_baselines.common.schedules import get_schedule_fn
 from stable_baselines.deepq.replay_buffer import ReplayBuffer
-from stable_baselines.ppo2.ppo2 import safe_mean, get_schedule_fn
 from stable_baselines.sac.policies import SACPolicy
 from stable_baselines import logger
